+++ config/schedule.rb
@@ -7,3 +7,8 @@
 every '0 * * * *' do
   runner 'SchedulerJob.perform_later'
 end
+
+# Clear expired users sessions
+every 1.day do
+  runner 'Users::Sessions::RemoveExpiredJob.perform_later'
+end