+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - save user session in cookies
+- removing expired users sessions
 
 ### Modified
 - change price for transfer in basketball